requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
attrs==23.1.0
azure-core==1.29.6
azure-identity==1.15.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
microsoft-kiota-abstractions==1.0.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.2.0
microsoft-kiota-serialization-json==1.0.0
microsoft-kiota-serialization-text==1.0.0
msal==1.26.0
msal-extensions==1.1.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0
multidict==6.0.4
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
packaging==23.2
pendulum==2.1.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytzdata==2020.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
std-uritemplate==0.0.49
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
azure-functions
async-timeout==4.0.2
azure-cosmos==4.5.0
click==8.1.5
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fastapi==0.110.0
fastapi-sessions==0.3.2
fonttools==4.41.0
future==0.18.3
httptools==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
pydantic==1.10.11
pyparsing==3.0.9
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
starlette==0.27.0
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
jwcrypto
openai
python-jose
semantic-kernel
fuzzywuzzy
python-Levenshtein